pipe: Reorganize files
