+++ pipeline/risteys_pipeline/finngen/surv_select_endpoint_pairs.py
