+++ pipeline/risteys_pipeline/utils/utils.py
@@ -1,9 +1,7 @@
 """Utils functions"""
 from contextlib import contextmanager
-
 import pandas as pd
-
-from risteys_pipeline.log import logger
+from risteys_pipeline.utils.log import logger
 
 
 DAYS_IN_YEAR = 365.25
