+++ pipeline/test/test_distributions.py
@@ -1,5 +1,5 @@
 import pandas as pd
-from risteys_pipeline.distributions import green_distribution
+from risteys_pipeline.run_distributions import green_distribution
 
 
 def test_green_distribution_early_onset():
